fix: SQLite 호환 — BigInteger→Integer PK + 이틀치 수집 시뮬레이션 테스트
- models: BigInteger primary key → Integer (SQLite autoincrement 호환)
- test_auto_collection.py: 48시간(16회) 자동 수집 시뮬레이션 테스트
  - 수집→NLP분석→DB저장→중복체크→이력기록 전체 파이프라인 검증
  - 48건 컨텐츠, 144건 댓글, 감성/키워드/중요도 전수 분석 확인

+++ backend/app/models/comment.py
@@ -25,11 +25,11 @@ class Comment(Base):
     __tablename__ = "comments"
 
     # 기본 키
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     # 소속 콘텐츠 외래 키
     content_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("contents.id", ondelete="CASCADE"),
         nullable=False,
         index=True,
